+++ apps/sim/lib/workflows/executor/pause-persistence.ts
@@ -1,9 +1,6 @@
 import { createLogger } from '@sim/logger'
-import { toError } from '@sim/utils/errors'
-import {
-  describeError,
-  isRetryableInfrastructureError,
-} from '@/lib/core/errors/retryable-infrastructure'
+import { describeError, toError } from '@sim/utils/errors'
+import { isRetryableInfrastructureError } from '@/lib/core/errors/retryable-infrastructure'
 import type { LoggingSession } from '@/lib/logs/execution/logging-session'
 import { PauseResumeManager } from '@/lib/workflows/executor/human-in-the-loop-manager'
 import type { ExecutionResult } from '@/executor/types'
